requirements.txt
Flask==1.1.1
google-cloud-datastore==1.7.3
google-api-core==1.22.1
google-cloud-bigquery==0.22.1